refactor(blue-migration-sdk): rename abis files
