+++ packages/migration-sdk-viem/src/fetchers/compoundV2/compoundV2.helpers.ts
@@ -2,7 +2,7 @@ import { ChainId, ChainUtils, MathLib } from "@morpho-org/blue-sdk";
 import type { FetchParameters } from "@morpho-org/blue-sdk-viem";
 import type { Address, Client } from "viem";
 import { getBlock, getChainId, readContract } from "viem/actions";
-import { cErc20Abi, mErc20Abi } from "../../abis/compoundV2.abis.js";
+import { cErc20Abi, mErc20Abi } from "../../abis/compoundV2.js";
 
 export interface CompoundV2MarketState {
   lastExchangeRate: bigint;
